ci: initial Packit configuration
Enable a basic set of Packit jobs to build util-linux RPMs from a
RHEL9/C9S specfile - on F34 for now, until we have public C9S images
available in Copr.

Some comments to the sed magic in the yaml configuration file:

1) We have to create a custom tarball by our own means, since the
   default Packit action doesn't include the necessary .tarball-version
   file. This wouldn't be necessary if we could use the `make distcheck`
   target, but the default builder chroot is missing some dependencies
   (like bison), so we have to fallback to `git archive`.

2) Consequently, we need to call autogen.sh from the specfile (normally
   we'd call it before `make distcheck` and already have a configure
   binary in the tarball)

3) Also consequently, explicitly run the `check` RPM phase during
   rpmbuild (by setting `_with_check 1`). This would be, also, normally
   executed by `make distcheck`.

As requested, run the checks on push events into the rhel-9 branch
instead of for each pull request.

This workflow can be later extended to send the built RPMs to the
Testing Farm (TFT)[0] for a more extensive testing.

[0] https://packit.dev/docs/testing-farm/